Updated Psudocode
* Removed unneeded datasheets
* Added new datasheets we didn’t have before
* Changed the location of the .X project so that it’s not nested inside
a useless folder
* Changed the PIC used in the .X project properties
* Added a lot of comments to the main.c file, including reworking the
program flow psudocode, adding interrupts and other functions I think
will be useful, a pin diagram, and other things.

I had to comb the datasheet for a good amount of time to find a lot of
the new data I added, and while I feel that most of it is accurate, it
could be off, so we need to verify these fundamental ideas.